requirements.txt
aerich==0.5.8
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
ddlparse==1.10.0
dictdiffer==0.9.0
docopt==0.6.2
ecdsa==0.17.0
fastapi==0.70.0
gunicorn==20.1.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
passlib==1.7.4
pipreqs==0.4.11
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyJWT==1.7.1
pyparsing==3.0.6
pypika-tortoise==0.1.1
python-decouple==3.3
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
websocket-client==1.2.1
websockets==10.1
yarg==0.1.9
numpy==1.22.2
geopy==2.2.0
pandas==1.4.1
scikit-learn==1.0.2
